+++ lib/paths.sh
@@ -73,6 +73,7 @@ gps_apply_paths() {
 	GPS_MESH_DIR="${GPS_ETC}/mesh"
 	GPS_MESH_PEERS="${GPS_MESH_DIR}/peers.json"
 	GPS_MESH_TOKEN_FILE="${GPS_MESH_DIR}/token"
+	GPS_MESH_JOIN_CMD="${GPS_MESH_DIR}/join.cmd"
 	GPS_MESH_TLS_CERT="${GPS_MESH_DIR}/master-tls.pem"
 	GPS_MESH_TLS_KEY="${GPS_MESH_DIR}/master-tls.key"
 	GPS_MESH_TLS_FP="${GPS_MESH_DIR}/master-tls.fp"